backend/requirements.txt
fastapi
uvicorn
langchain
chromadb
sentence-transformers
python-multipart
even-glasses
google-generativeai
pydantic-settings
pypdf
langchain-community
langchain-text-splitters
langchain-chroma